Store queue sizes in u_int, not uint16_t.
65536 is a valid queue size but wraps down to 0 in a uint16_t causing
confusion.

+++ lib/libnvmf/internal.h
@@ -87,7 +87,7 @@ struct nvmf_qpair {
 	 * Queue sizes.  This assumes the same size for both the
 	 * completion and submission queues within a pair.
 	 */
-	uint16_t nq_qsize;
+	u_int	nq_qsize;
 
 	/* Flow control management for submission queues. */
 	bool nq_flow_control;
